# Extract faster-babel-loader-for-macro to speed up build time