--- a/packages/cli/package.json
+++ b/packages/cli/package.json
@@ -71,6 +71,7 @@
     "@types/yargs": "^15.0.5"
   },
   "peerDependencies": {
-    "@goji/core": "*"
+    "@goji/core": "*",
+    "source-map": "*"
   }
 }