--- a/packages/cli/src/config/webpack.config.ts
+++ b/packages/cli/src/config/webpack.config.ts
@@ -112,7 +112,8 @@ export const getWebpackConfig = ({
           use: [
             // support `@goji/macro`
             {
-              loader: require.resolve('babel-loader'),
+              // loader: require.resolve('babel-loader'),
+              loader: require.resolve('./fasterBabelLoaderForMacro'),
               options: {
                 plugins: [[require.resolve('babel-plugin-macros'), { gojiMacro: { target } }]],
               },