To support `@goji/macro` feature, the Webpack always create a new instance of `babel` in each call of `babel-loader`.
But in fact, we could only parse files that include `import from macro` statement. So I create a `faster-babel-loader-for-macro` that check whether the file match [`/[./]macro(\.c?js)?$/`](https://github.com/kentcdodds/babel-plugin-macros/blob/18d79d3ac3c2975565e7897d2233cf498f15ffb5/src/index.js#L5) regex before parse them.

```tsx
import { registerPluginComponent } from '@goji/macro'; // should parse this file
```

This PR reduced 25% Webpack bundling time.

|     | Before | After |
| --- | ------ | ----- |
| 1   | 13.3   | 10.26 |
| 2   | 12.65  | 9.63  |
| 3   | 12.82  | 9.5   |
| 4   | 12.76  | 9.21  |
| 5   | 12.67  | 9.52  |
| Avg | 12.84  | 9.624 |